requirements.txt
dimod==0.12.13
dwave_cloud_client==0.11.1
dwave_drivers==0.4.4
dwave_inspector==0.4.3
dwave_inspectorapp==0.3.1
dwave_preprocessing==0.6.4
dwave_samplers==1.2.0
dwave_system==1.21.0
numpy==1.26.2
